Allow interchangeable use of CGFloat and Double types
SE-0307

+++ Shared/Modules/ChartsExtensions/Sources/ChartsExtensions/PieChartView+Extension.swift
@@ -63,7 +63,7 @@ public extension PieChartView {
                 legend.verticalAlignment   = .bottom
         }
         legend.horizontalAlignment = .left
-        legend.formSize = CGFloat(12.0)
+        legend.formSize = 12.0
         
         //: ### Description
         self.chartDescription?.enabled = descriptionEnabled
